Further break up behavior to build up with composition.
